WIP: move some api scripts to unity for testing
