+++ src/ValheimRAFT/AssetBundle/AssetBundle.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 3048082280
+CRC: 3482909097
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
