+++ src/ValheimRAFT/AssetBundle/valheim-vehicles-shared.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 2879305702
+CRC: 3212790301
 Hashes:
   AssetFileHash:
     serializedVersion: 2
